Collapsabel toolbar in home screen (#297)
* Added every category from OpenTDBApi

* Collapsing toolbar in progress

* Collapsing toolbar some progress

* SearchBar in progress

* Gird header is finished for now

* Fixed symbols problem in answers

* refactor OpenTdbRepoImpl

* small cleanup
