+++ testomania/src/main/java/com/earth/testomania/common/custom_ui_components/UITools.kt
@@ -1,0 +1,7 @@
+package com.earth.testomania.common.custom_ui_components
+
+import androidx.compose.material.ScaffoldState
+
+fun dismissSnackbar(scaffoldState: ScaffoldState?) {
+    scaffoldState?.snackbarHostState?.currentSnackbarData?.dismiss()
+}
